refactor(bindings): extract Top2Vec + BERTopic into src/python/embedding_cluster.rs (#385)
Second incremental split of src/python/mod.rs. Move the two embedding-clustering
models — Top2Vec and BERTopic — into a new src/python/embedding_cluster.rs. They
share the class-based TF-IDF topic_word path and the `umap_notice` reducer guard
(used only by these two), which moves with them; their save-structs
(Top2VecState/BertopicState) were already co-located and move too. Shared helpers,
type aliases, and save-format tags stay in mod.rs, reached via `use super::*`.

No public API change: both classes are still registered in the same #[pymodule]
fn (`use embedding_cluster::{BERTopic, Top2Vec}`) and imported as topica.Top2Vec /
topica.BERTopic. Update the #381 IMPL map bindings to point at the new module and
regenerate the model map.

mod.rs: 17658 -> 16414 lines. Gates green: cargo check (default + python), clippy
-D warnings, fmt, preflight, and the full pytest suite (3321 passed, incl. cluster
gold/parity, embedding save/load + transform, conformance, registry).
