+++ python/topica/registry.py
@@ -281,8 +281,8 @@ IMPL: dict[str, ImplInfo] = {
     "DETM": _i("src/detm.rs", "src/python/mod.rs", "embedding VAE + LSTM q(eta) (etm_vae.rs)", "", "parity/detm_gold.py"),
     "HLDA": _i("src/hlda.rs", "src/python/hierarchical.rs", "nested-CRP collapsed Gibbs", "", "parity/hlda_gold.py"),
     "PA": _i("src/pa.rs", "src/python/hierarchical.rs", "collapsed Gibbs over a topic DAG", "", "parity/pa_gold.py"),
-    "BERTopic": _i("src/bertopic.rs", "src/python/mod.rs", "embedding clustering (cluster.rs, reduce.rs, represent.rs)", "embeddings", "parity/bertopic_gold.py"),
-    "Top2Vec": _i("src/top2vec.rs", "src/python/mod.rs", "embedding clustering (cluster.rs, reduce.rs)", "embeddings", "parity/top2vec_gold.py, parity/top2vec_compare.py"),
+    "BERTopic": _i("src/bertopic.rs", "src/python/embedding_cluster.rs", "embedding clustering (cluster.rs, reduce.rs, represent.rs)", "embeddings", "parity/bertopic_gold.py"),
+    "Top2Vec": _i("src/top2vec.rs", "src/python/embedding_cluster.rs", "embedding clustering (cluster.rs, reduce.rs)", "embeddings", "parity/top2vec_gold.py, parity/top2vec_compare.py"),
     "ETM": _i("src/etm.rs", "src/python/mod.rs", "variational EM over word embeddings (ctm.rs)", "", "parity/etm_gold.py"),
     "IdealPointTM": _i("src/idealpoint.rs", "src/python/idealpoint.rs", "variational EM + ideal-point head", "", "tests/test_idealpoint.py, tests/test_idealpoint_counts.py"),
     "Wordfish": _i("src/wordfish.rs", "src/python/wordfish.rs", "Poisson-scaling EM", "", "parity/wordfish_r_compare.py, tests/test_wordfish.py"),
